Changed a bunch of things
I decided to rewrite most of the SetPos and SetSize bits so they would work with docking instead, which also makes the main panel able to be resized now, so I enabled resizing on line 17 because of this. I also changed the paint functions a little, and moved some of the code around, and I removed the scrollpanel that houses the two DListViews since DListViews have their own scrollpanel.